Unique creature's status was updated. A bug about patron index was fixed.
git-svn-id: http://svn.osdn.net/svnroot/deeangband@1299 89dd94fa-70ca-40ae-8747-2094edff379b

+++ v0_1_0/trunk/src/monster2.c
@@ -3702,7 +3702,7 @@ int create_creature(creature_type *creature_ptr, int species_idx, int monster_eg
 	creature_ptr->race_idx1 = (s16b)species_ptr->race_idx1;
 	creature_ptr->race_idx2 = (s16b)species_ptr->race_idx2;
 	creature_ptr->cls_idx = (byte)species_ptr->cls_idx;
-	creature_ptr->patron_idx = (byte)species_ptr->patron_idx;
+	creature_ptr->patron_idx = species_ptr->patron_idx;
 	creature_ptr->chara_idx = (byte)species_ptr->chara_idx;
 	creature_ptr->ap_species_idx = species_idx;
